Fix Update Version step in push GHA
We're seeing a 403 when attempting to push with the create-pull-
request action. I believe this is being caused by the fact that the
checkout is being done using the default action token, but the push
is trying to use the onf-bot's PAT. Using the same token for both
should fix the error.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>